Copy assets file into volume at runtime
Due to https://github.com/docker/compose/issues/11953 updating to a new
assets image doesn't guarantee that the new contents of the exposed
volume will be used.

Therefore the files are now copied into the volume at runtime.
This ensures we're always using up to date files.

To make sure dependent containers don't use partial state we now also
have a healthcheck to ensure that the copying is done before the contents
are being used.